Kuvagalleria/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.12.3
blinker==1.4
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.5
command-not-found==0.3
crispy-bootstrap5==0.7
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.0.6
django-appconf==1.0.5
django-bootstrap4==24.1
django-crispy-forms==1.14.0
django-imagekit==4.1.0
ExifRead==3.0.0
filesystemlib==0.1.0
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
Panda3D==1.10.14
pilkit==2.0
Pillow==9.2.0
pycparser==2.21
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sorl-thumbnail==12.10.0
soupsieve==2.5
sqlparse==0.4.2
systemd-python==234
tzdata==2022.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
wadllib==1.3.6
zipp==1.0.0
